double precision

+++ apps/bridson-density/advection.hpp
@@ -11,7 +11,7 @@ public:
     /// @param u Velocity field along the x axis.
     /// @param v Velocity field along the y axis.
     /// @return Grid at the updated state.
-    void operator()(const f32 dt);
+    void operator()(const f64 dt);
 
     /// @brief Swaps the back buffer grid with mQ. Necessary self-advection.
     void swap();
@@ -24,14 +24,14 @@ private:
 
     /// @brief Determines the initial position of an imaginary particle with
     /// current position pos integrated back in time by dt.
-    Vector2D backtrace(const Vector2D& grid_pos, const f32 dt) const;
+    Vector2D backtrace(const Vector2D& grid_pos, const f64 dt) const;
 
     /// @brief Integrates back in time using Euler.
-    Vector2D euler(const Vector2D& grid_pos, const f32 dt) const;
+    Vector2D euler(const Vector2D& grid_pos, const f64 dt) const;
 
     /// @brief Integrates back in time with 2nd order Runge-Kutta.
-    Vector2D RK2(const Vector2D& grid_pos, const f32 dt) const;
+    Vector2D RK2(const Vector2D& grid_pos, const f64 dt) const;
 
     /// @brief Integrates back in time with 3nd order Runge-Kutta.
-    Vector2D RK3(const Vector2D& grid_pos, const f32 dt) const;
+    Vector2D RK3(const Vector2D& grid_pos, const f64 dt) const;
 };
