+++ apps/bridson-density/config.hpp
@@ -5,9 +5,9 @@
 struct Config {
     Size rows;
     Size cols;
-    f32 cellSize;
-    f32 timestep;
-    f32 density;
+    f64 cellSize;
+    f64 timestep;
+    f64 density;
     bool saveFrames;
 
     static Config loadFromJson(const std::string& path);
